move db classe to tables companion objects
