--- db/src/main/scala/slick/data/BlockSet.scala
@@ -1,6 +1,0 @@
-package slick.data
-
-final case class BlockSet(
-  id: Long,         // primary key
-  hash: Array[Byte],// strong hash of a hashes of blocks in a set (global identifier)
-)
