--- db/src/main/scala/slick/data/BondSet.scala
@@ -1,6 +1,0 @@
-package slick.data
-
-final case class BondSet(
-  id: Long,         // primary key
-  hash: Array[Byte],// strong hash of a bonds set content (global identifier)
-)
