--- db/src/main/scala/slick/data/DeploySet.scala
@@ -1,6 +1,0 @@
-package slick.data
-
-final case class DeploySet(
-  id: Long,         // primary key
-  hash: Array[Byte],// strong hash of a signatures of deploys in a set (global identifier)
-)
