--- db/src/main/scala/slick/data/Deployer.scala
@@ -1,6 +1,0 @@
-package slick.data
-
-final case class Deployer(
-  id: Long,           // primary key
-  pubKey: Array[Byte],// public key of a deployer
-)
