--- db/src/main/scala/slick/data/Shard.scala
@@ -1,6 +1,0 @@
-package slick.data
-
-final case class Shard(
-  id: Long,    // primary key
-  name: String,// shard name
-)
