--- db/src/main/scala/slick/data/Validator.scala
@@ -1,6 +1,0 @@
-package slick.data
-
-final case class Validator(
-  id: Long,           // primary key
-  pubKey: Array[Byte],// validator public key
-)
