+++ db/src/main/scala/slick/tables/TableBlockSet.scala
@@ -1,12 +1,19 @@
 package slick.tables
 
-import slick.data.BlockSet
 import slick.jdbc.PostgresProfile.api.*
 import slick.lifted.ProvenShape
+import slick.tables.TableBlockSet.BlockSet
 
 class TableBlockSet(tag: Tag) extends Table[BlockSet](tag, "block_set") {
   def id: Rep[Long]          = column[Long]("id", O.PrimaryKey, O.AutoInc)
   def hash: Rep[Array[Byte]] = column[Array[Byte]]("hash", O.Unique)
 
   def * : ProvenShape[BlockSet] = (id, hash).mapTo[BlockSet]
+}
+
+object TableBlockSet {
+  final case class BlockSet(
+    id: Long,         // primary key
+    hash: Array[Byte],// strong hash of a hashes of blocks in a set (global identifier)
+  )
 }
