+++ db/src/main/scala/slick/tables/TableBond.scala
@@ -11,3 +11,11 @@ class TableBond(tag: Tag) extends Table[Bond](tag, "bond") {
 
   def * : ProvenShape[Bond] = (validatorId, stake).mapTo[Bond]
 }
+
+object TableBond {
+  final case class Bond(
+    id: Long,          // primary key
+    validatorId: Long, // pointer to a validator
+    stake: Long,       // stake of a validator
+  )
+}
