+++ db/src/main/scala/slick/tables/TableBondSet.scala
@@ -1,12 +1,19 @@
 package slick.tables
 
-import slick.data.BondSet
 import slick.jdbc.PostgresProfile.api.*
 import slick.lifted.ProvenShape
+import slick.tables.TableBondSet.BondSet
 
 class TableBondSet(tag: Tag) extends Table[BondSet](tag, "bond_set") {
   def id: Rep[Long]          = column[Long]("id", O.PrimaryKey, O.AutoInc)
   def hash: Rep[Array[Byte]] = column[Array[Byte]]("hash", O.Unique)
 
   def * : ProvenShape[BondSet] = (id, hash).mapTo[BondSet]
+}
+
+object TableBondSet {
+  final case class BondSet(
+    id: Long,         // primary key
+    hash: Array[Byte],// strong hash of a bonds set content (global identifier)
+  )
 }
