+++ db/src/main/scala/slick/tables/TableBondSetBind.scala
@@ -1,8 +1,8 @@
 package slick.tables
 
-import slick.data.BondSetBind
 import slick.jdbc.PostgresProfile.api.*
 import slick.lifted.ProvenShape
+import slick.tables.TableBondSetBind.BondSetBind
 
 class TableBondSetBind(tag: Tag) extends Table[BondSetBind](tag, "bond_set_bind") {
   def bondSetId: Rep[Long] = column[Long]("bond_set_id")
@@ -14,4 +14,11 @@ class TableBondSetBind(tag: Tag) extends Table[BondSetBind](tag, "bond_set_bind"
   def idx = index("idx_bond_set_bind", bondSetId, unique = false)
 
   def * : ProvenShape[BondSetBind] = (bondSetId, bondId).mapTo[BondSetBind]
+}
+
+object TableBondSetBind {
+  final case class BondSetBind(
+    bondSetId: Long, // pointer to bondSet
+    bondId: Long,    // pointer to bond
+  )
 }
