+++ db/src/main/scala/slick/tables/TableDeploySet.scala
@@ -1,12 +1,19 @@
 package slick.tables
 
-import slick.data.DeploySet
 import slick.jdbc.PostgresProfile.api.*
 import slick.lifted.ProvenShape
+import slick.tables.TableDeploySet.DeploySet
 
 class TableDeploySet(tag: Tag) extends Table[DeploySet](tag, "deploy_set") {
   def id: Rep[Long]          = column[Long]("id", O.PrimaryKey, O.AutoInc)
   def hash: Rep[Array[Byte]] = column[Array[Byte]]("hash", O.Unique)
 
   def * : ProvenShape[DeploySet] = (id, hash).mapTo[DeploySet]
+}
+
+object TableDeploySet {
+  final case class DeploySet(
+    id: Long,         // primary key
+    hash: Array[Byte],// strong hash of a signatures of deploys in a set (global identifier)
+  )
 }
