+++ db/src/main/scala/slick/tables/TableDeployer.scala
@@ -1,8 +1,8 @@
 package slick.tables
 
-import slick.data.Deployer
 import slick.jdbc.PostgresProfile.api.*
 import slick.lifted.ProvenShape
+import slick.tables.TableDeployer.Deployer
 
 class TableDeployer(tag: Tag) extends Table[Deployer](tag, "deployer") {
   def id: Rep[Long]            = column[Long]("id", O.PrimaryKey, O.AutoInc)
@@ -11,4 +11,11 @@ class TableDeployer(tag: Tag) extends Table[Deployer](tag, "deployer") {
   def idx = index("idx_deployer", pubKey, unique = true)
 
   def * : ProvenShape[Deployer] = (id, pubKey).mapTo[Deployer]
+}
+
+object TableDeployer {
+  final case class Deployer(
+    id: Long,           // primary key
+    pubKey: Array[Byte],// public key of a deployer
+  )
 }
