+++ db/src/main/scala/slick/tables/TableValidator.scala
@@ -1,8 +1,8 @@
 package slick.tables
 
-import slick.data.Validator
 import slick.jdbc.PostgresProfile.api.*
 import slick.lifted.ProvenShape
+import slick.tables.TableValidator.Validator
 
 class TableValidator(tag: Tag) extends Table[Validator](tag, "validator") {
   def id: Rep[Long]            = column[Long]("id", O.PrimaryKey, O.AutoInc)
@@ -11,4 +11,11 @@ class TableValidator(tag: Tag) extends Table[Validator](tag, "validator") {
   def idx = index("idx_validator", pubKey, unique = true)
 
   def * : ProvenShape[Validator] = (id, pubKey).mapTo[Validator]
+}
+
+object TableValidator {
+  final case class Validator(
+    id: Long,           // primary key
+    pubKey: Array[Byte],// validator public key
+  )
 }
